backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
brotlipy==0.5.1
cffi==1.8.2               # via brotlipy
django-environ==0.4.0
django-extensions==1.7.4
django-filter==0.15.3
Django==1.10.2            # via django-environ
djangorestframework==3.4.6
grako==3.14.0
gunicorn==19.6.0
psycopg2==2.6.2
pycparser==2.14           # via cffi
six==1.10.0               # via django-environ, django-extensions
Werkzeug==0.11.11
whitenoise==3.2.1
